usb: otg: add cherryview support
This adds CHT phy mux and xhci capabilities for Aero dev kit.
Portions of the original code have been removed to implement
basic features only.

This code is not intended to be submitted to LKML, cleanup and
adaptation is needed for latest kernels. In future, this patch
should be reverted in favor of a driver that has been reviewed
and accepted in LKML.

As such, this contains multiple logical blocks to keep changes
contained in a single patch. Since it is a big squash patch,
I think it is worth mentioning the people that have contributed
to this code:

    David A Cohen <david.a.cohen@intel.com>
    Dawid Manikowski <dawid.manikowski@intel.com>
    Du, Changbin <changbin.du@intel.com>
    Fei Yang <fei.yang@intel.com>
    fu jin <fu.jin@intel.com>
    Hetian Fang <hetian.fang@intel.com>
    Jiebing Li <jiebing.li@intel.com>
    Konrad Leszczynski <konrad.leszczynski@intel.com>
    Liu Linhong <linhong.liu@intel.com>
    Lu Baolu <baolu.lu@linux.intel.com>
    Matt Ranostay <matt.ranostay@intel.com>
    Miroslaw Swiader <miroslaw.swiader@intel.com>
    mwalews <marcin.walewski@intel.com>
    Robert Dobrowolski <robert.dobrowolski@intel.com>
    Wu, Hao <hao.wu@intel.com>
    xiao jin <jin.xiao@intel.com>
    yangzhon <yang.zhong@intel.com>
    Zhang Yanmin <yanmin.zhang@intel.com>
    Zhuang Jin Can <jin.can.zhuang@intel.com>

Signed-off-by: Omar Ramirez Luna <omar.r.ramirez.luna@intel.com>