pmic: intel: port whiskey cove driver
Adds Whiskey Cove PMIC driver and features related to USB OTG
detection, charger and temp have been removed along with portions
of the original code to implement basic features only.

This code is not intended to be submitted to LKML, cleanup and
adaptation is needed for latest kernels. In future, this patch
should be reverted in favor of a driver that has been reviewed
and accepted in LKML.

As such, this contains multiple logical blocks to keep changes
contained in a single patch. Since it is a big squash patch,
I think it is worth mentioning the people that have contributed
to this code:

    Aaron Lu <aaron.lu@intel.com>
    Albin B <albin.bala.krishnan@intel.com>
    Ananth Krishna R <ananth.krishna.r@intel.com>
    Bidipta Samadder <bidiptax.samadder@intel.com>
    Biju Chakkachattil <bijux.chakkachattil@intel.com>
    Catalin Popescu <catalin.popescu@intel.com>
    Ding Yunliang <yunliang.ding@intel.com>
    Fei Yang <fei.yang@intel.com>
    Jenny TC <jenny.tc@intel.com>
    John Fredriksson <john.fredriksson@intel.com>
    ola.lilja <ola.lilja@intel.com>
    Pat Noziska <patrick.j.noziska@intel.com>
    qipeng.zha <qipeng.zha@intel.com>
    Rakesh Ranjan Jha <rakesh.r.jha@intel.com>
    raosrini <srinidhi.rao@intel.com>
    Rapaka Naveen <naveen.rapaka@intel.com>
    Richard A. Griffiths <richard.a.griffiths@intel.com>
    R, Kannappan <r.kannappan@intel.com>
    srihars1 <sriharsha.govinda.raj@intel.com>
    Venkataramana Kotakonda <venkataramana.kotakonda@intel.com>
    Vineesh k k <vineesh.k.k@intel.com>
    ysiyer <yegnesh.s.iyer@intel.com>
    Yuan Juntao <juntao.yuan@intel.com>
    Zhang Yanmin <yanmin.zhang@intel.com>

Signed-off-by: Omar Ramirez Luna <omar.r.ramirez.luna@intel.com>